Relocated 'colors' package in root path and renamed it to 'ui'
