+++ jplotter/src/main/java/hageldave/jplotter/renderers/CoordSysRenderer.java
@@ -9,9 +9,9 @@ import hageldave.jplotter.interaction.CoordinateViewListener;
 import hageldave.jplotter.renderables.Legend;
 import hageldave.jplotter.renderables.Lines;
 import hageldave.jplotter.renderables.Text;
-import hageldave.jplotter.renderers.colors.ColorProvider;
-import hageldave.jplotter.renderers.colors.schemes.ColorScheme;
 import hageldave.jplotter.svg.SVGUtils;
+import hageldave.jplotter.ui.ColorProvider;
+import hageldave.jplotter.ui.schemes.ColorScheme;
 import hageldave.jplotter.util.Annotations.GLCoordinates;
 import hageldave.jplotter.util.Pair;
 import hageldave.jplotter.util.PointeredPoint2D;
